File sturcture updated, much more dynamic. Id prop solved for dynamic language ids on default button class. Probs all I have for this epic Friday night.
