+++ error_log
@@ -1,0 +1,16 @@
+[06-Apr-2019 04:38:29 UTC] PHP Warning:  fopen(../languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 40
+[06-Apr-2019 04:40:24 UTC] PHP Warning:  fopen(../languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 40
+[06-Apr-2019 04:41:47 UTC] PHP Warning:  fopen(../languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 40
+[06-Apr-2019 04:45:43 UTC] PHP Warning:  fopen(languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 39
+[06-Apr-2019 04:46:09 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:46:09 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
+[06-Apr-2019 04:47:19 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:47:19 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
+[06-Apr-2019 04:47:32 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:47:32 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
+[06-Apr-2019 04:48:36 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:48:36 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
+[06-Apr-2019 04:49:18 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:49:18 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
+[06-Apr-2019 04:50:13 UTC] PHP Warning:  file_get_contents(/home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/languages/supported-languages.json): failed to open stream: No such file or directory in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 52
+[06-Apr-2019 04:50:13 UTC] PHP Warning:  Invalid argument supplied for foreach() in /home/thewebbranch/public_html/git/Google-Translate-PHP-URL/translator_class_library/translate_init.php on line 55
